% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Huarpa.R
\docType{data}
\name{Huarpa}
\alias{Huarpa}
\title{A data set of precipitation historical series of stations of Huarpa basin (Peru).}
\format{
A dataset of precipitation historical monthly series.
\describe{
  \item{Fecha}{first column by months, from 01-01-1964 to 31-12-2016}
  ...
}
}
\source{
data downloaded from SENAMHI (Peru), these data being distributed without any warranty.
}
\usage{
Huarpa
}
\description{
precipitation historical monthly data of stations of Huarpa basin (SENAMHI, Peru).
}
\keyword{datasets}
