% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Huarpa.stations.R
\docType{data}
\name{Huarpa.stations}
\alias{Huarpa.stations}
\title{Stations information of basin Huarpa (SENAMHI, Peru).}
\format{
a dataframe with information of 22 meteorological stations.
\describe{
  \item{Estacion}{Stations names}
  \item{Lat}{Latitude, in decimal degrees}
  \item{Lon}{Longitude, in decimal degrees}
  \item{Alt}{Altitude, in msnm.}
  ...
}
}
\source{
data downloaded from SENAMHI (Peru), these data being distributed without any warranty.
}
\usage{
Huarpa.stations
}
\description{
dataframe with information of 22 meteorological stations. (SENAMHI,Peru)
}
\keyword{datasets}
