% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyavg.R
\name{monthlyavg}
\alias{monthlyavg}
\title{function for processing of pisco daily values to a monthly average.}
\usage{
monthlyavg(x, param = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}
}
\description{
function for processing of pisco daily values to a monthly average, apply to precipitation,
temperature and evapotranspiration.
}
\author{
Geomar Perales Apaico
}
