% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisco2monthly.R
\name{pisco2monthly}
\alias{pisco2monthly}
\title{function for processing of pisco daily values to a monthly serie.}
\usage{
pisco2monthly(x, variable = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily values.}

\item{variable}{OPTIONAL, default value is sum for precipitation, use mean for temperature.}
}
\description{
function for processing of pisco daily values to a monthly serie, apply to precipitation and
temperature.
}
