% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscod.eo.R
\name{piscod.eo}
\alias{piscod.eo}
\title{values extraction of PISCO evapotranspiration daily climatic databases}
\usage{
piscod.eo(x)
}
\arguments{
\item{x}{a dataframe with PISCO files folder (netCDF format) of evapotranspiration, longitude and latitude of station.}
}
\description{
function for extract values of PISCO evapotranspiration daily climatic databases,
PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\examples{

library(RPisco)

PISCO.folder <- "D:/0-BD-Pisco/0-Evp"
latitude <- -76.11
longitude <- -13.11
x <-  data.frame(PISCO.folder, latitude, longitude)
piscod.eo(x)

}
\author{
Geomar Perales Apaico
}
