% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscom.R
\name{piscom}
\alias{piscom}
\title{values extraction of PISCO monthly climatic databases from a station}
\usage{
piscom(x, obj = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO file name (in netCDF format), longitude and latitude of station.}

\item{obj}{OPTIONAL, default parameter is NULL, use this parameter for obtain data in
column or matrix. Use col for obtain data column or matrix for obtain data matrix.}
}
\description{
function for extract values of PISCO monthly climatic databases, PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\examples{

library(RPisco)

PISCO.file <- file.path("D:", "PISCOm_pp_2.0.nc")
latitude <- -76.11
longitude <- -13.11
x <-  data.frame(PISCO.file, latitude, longitude)
piscom(x)

}
\author{
Geomar Perales Apaico
}
