% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscomgroup.R
\name{piscomgroup}
\alias{piscomgroup}
\title{values extraction of PISCO monthly climatic databases from a stations group}
\usage{
piscomgroup(x)
}
\arguments{
\item{x}{A dataframe containing the PISCO file name (in netCDF format), longitude and latitude from a stations group.}
}
\description{
function for extract values of PISCO monthly climatic databases from a stations group, PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\details{
To use RPisco, Download PISCO climatic databases from IRI/LDEO Climate Data Library.
visit: \href{https://iridl.ldeo.columbia.edu/SOURCES/.SENAMHI/.HSR/.PISCO/}{IRI/LDEO-SENAMHI-PISCO}
}
\examples{

library(RPisco)

PISCO.file <- file.path("D:", "PISCOm_pp_2.0.nc")
station <- c("Ayacucho", "Coracora")
latitude <- c(-74.23, -73.78)
longitude <- c(-13.18, -15.01)
x <-  data.frame(PISCO.file, station, latitude, longitude)
piscomgroup(x)

}
\author{
Geomar Perales Apaico
}
