% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyavg.R
\name{yearlyavg}
\alias{yearlyavg}
\title{Function for processing PISCO daily/monthly values into a yearly average}
\usage{
yearlyavg(x, param = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily/monthly values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}
}
\description{
Function for processing PISCO daily/monthly values into a yearly average, apply to precipitation,
temperature and evapotranspiration.
}
\author{
Geomar Perales Apaico
}
