% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyavg.R
\name{monthlyavg}
\alias{monthlyavg}
\title{function for processing of pisco daily values to a monthly average.}
\usage{
monthlyavg(x, variable = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily values.}

\item{variable}{OPTIONAL, default value is sum for precipitation, use mean for temperature.}
}
\description{
function for processing of pisco daily values to a monthly average, apply to precipitation and
temperature.
}
