% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisco2annual.R
\name{pisco2annual}
\alias{pisco2annual}
\title{function for processing of pisco daily/monthly values to a annual serie.}
\usage{
pisco2annual(x, variable = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily/monthly values.}

\item{variable}{OPTIONAL, default value is sum for precipitation, use mean for temperature.}
}
\description{
function for processing of pisco daily/monthly values to a annual serie, apply to precipitation and
temperature.
}
