% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisco2annual.R
\name{pisco2annual}
\alias{pisco2annual}
\title{Function for processing PISCO daily/monthly values into an annual series}
\usage{
pisco2annual(x, param = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily/monthly values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}
}
\description{
Function for processing PISCO daily/monthly values into an annual series, apply to precipitation,
temperature and evapotranspiration.
}
\examples{

library(RPisco)

pisco.evp <- "D:/0-BD-Pisco/0-Evp"
latitude <- -76.11
longitude <- -13.11
x <-  data.frame(pisco.evp, latitude, longitude)
data.diaria.eo <- piscod.eo(x)
pisco2annual(data.diaria.eo)

}
\author{
Geomar Perales Apaico
}
