% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisco2monthly.R
\name{pisco2monthly}
\alias{pisco2monthly}
\title{function for processing of pisco daily values to a monthly serie.}
\usage{
pisco2monthly(x, param = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}
}
\description{
function for processing of pisco daily values to a monthly serie, apply to precipitation,
temperature and evapotranspiration.
}
\author{
Geomar Perales Apaico
}
