% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisco2monthly.R
\name{pisco2monthly}
\alias{pisco2monthly}
\title{Function for processing PISCO daily values into a monthly series}
\usage{
pisco2monthly(x, param = NULL, obj = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}

\item{obj}{OPTIONAL, default parameter is NULL, use this parameter for obtain data in
column or matrix. Use col for obtain data column or matrix for obtain data matrix.}
}
\description{
Function for processing PISCO daily values into a monthly series, apply to precipitation,
temperature and evapotranspiration.
}
\author{
Geomar Perales Apaico
}
