% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisco2monthly.R
\name{pisco2monthly}
\alias{pisco2monthly}
\title{Function for processing PISCO daily values into a monthly series}
\usage{
pisco2monthly(x, param = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}
}
\description{
Function for processing PISCO daily values into a monthly series, apply to precipitation,
temperature and evapotranspiration.
}
\examples{

library(RPisco)

pisco.ppd <- "D:/PISCOd_pp_2.0.nc"
latitude <- -76.11
longitude <- -13.11
x <-  data.frame(pisco.ppd, latitude, longitude)
data.diaria.pp <- piscod(x)
pisco2monthly(data.diaria.pp)

}
\author{
Geomar Perales Apaico
}
