% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{pisco_catalog}
\alias{pisco_catalog}
\title{Catalog of PISCO Products}
\usage{
pisco_catalog()
}
\value{
A \link[tibble:tibble]{tibble::tibble} containing metadata for all available datasets.
}
\description{
Lists the available PISCO datasets, metadata, temporal ranges,
resolution, official file names, and current local cache status.
}
\examples{
pisco_catalog()
}
