% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscod.eo.R
\name{piscod.eo}
\alias{piscod.eo}
\title{values extraction of PISCO evapotranspiration daily climatic databases}
\usage{
piscod.eo(x)
}
\arguments{
\item{x}{a dataframe with PISCO files folder (netCDF format) of evapotranspiration, longitude and latitude of station.}
}
\description{
function for extract values of PISCO evapotranspiration daily climatic databases,
PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\details{
To use piscod.eo, Download PISCO climatic databases from IRI/LDEO Climate Data Library.
visit: \href{https://figshare.com/articles/dataset/Reference_crop_evapotranspiration_PISCOeo_pm_/15215106?backTo=/collections/A_reference_evapotranspiration_gridded_database_based_on_FAO_Penman-Monteith_in_Peru_during_1981-2016/5633182}{PISCOeo_pm files from Figshare}
}
\examples{

library(RPisco)

PISCO.folder <- "D:/0-BD-Pisco/0-Evp"
latitude <- -74.23
longitude <- -13.18
x <-  data.frame(PISCO.folder, latitude, longitude)
piscod.eo(x)

}
\author{
Geomar Perales Apaico
}
