% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscod.R
\name{piscod}
\alias{piscod}
\title{values extraction of PISCO daily climatic databases from a station}
\usage{
piscod(x)
}
\arguments{
\item{x}{A dataframe containing the PISCO file name (in netCDF format), longitude and latitude of station.}
}
\description{
function for extract values of PISCO daily climatic databases, PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\details{
To use RPisco, Download PISCO climatic databases from IRI/LDEO Climate Data Library.
visit: \href{https://iridl.ldeo.columbia.edu/SOURCES/.SENAMHI/.HSR/.PISCO/}{IRI/LDEO-SENAMHI-PISCO}
}
\examples{

library(RPisco)

PISCO.file <- file.path("D:", "PISCOd_pp_2.0.nc")
latitude <- -74.23
longitude <- -13.18
x <-  data.frame(PISCO.file, latitude, longitude)
piscod(x)

}
\author{
Geomar Perales Apaico
}
