% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscodgroup.R
\name{piscodgroup}
\alias{piscodgroup}
\title{values extraction of PISCO daily climatic databases from a stations group.}
\usage{
piscodgroup(x, start = NULL, end = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO file name (netCDF format), longitude and latitude from stations group.}

\item{start}{OPTIONAL, start date of extraction period, default value is 1981-01-01.}

\item{end}{OPTIONAL, end date of extraction period, default value is 2016-12-31.}
}
\description{
function for extract values of PISCO daily climatic databases from a stations group, PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\examples{

library(RPisco)

PISCO.file <- file.path("D:", "PISCOd_pp_2.0.nc")
station <- "Ayacucho"
latitude <- -76.11
longitude <- -13.11
x <-  data.frame(station, latitude, longitude)
piscodgroup(PISCO.file, x)

}
\author{
Geomar Perales Apaico
}
