% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscom.R
\name{piscom}
\alias{piscom}
\title{values extraction of PISCO monthly data}
\usage{
piscom(x, obj = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO file name (in netCDF format), longitude and latitude of station.}

\item{obj}{OPTIONAL, default parameter is NULL, use this parameter for obtain data in
column or matrix. Use col for obtain data column or matrix for obtain data matrix.}
}
\description{
function for extract values of PISCO monthly data, PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\author{
Geomar Perales Apaico
}
