% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscom.R
\name{piscom}
\alias{piscom}
\title{values extraction of PISCO monthly climatic databases from a station}
\usage{
piscom(x)
}
\arguments{
\item{x}{A dataframe containing the PISCO file name (in netCDF format), longitude, and latitude of the station.}
}
\description{
function for extract values of PISCO monthly climatic databases, PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
\details{
To use Piscom, Download PISCO climatic databases from IRI/LDEO Climate Data Library.
visit: \href{https://iridl.ldeo.columbia.edu/SOURCES/.SENAMHI/.HSR/.PISCO/}{IRI/LDEO-SENAMHI-PISCO}
}
\examples{

library(RPisco)

PISCO.file <- file.path("D:", "PISCOm_pp_2.0.nc")
latitude <- -74.23
longitude <- -13.18
x <-  data.frame(PISCO.file, latitude, longitude)
piscom(x)

}
\author{
Geomar Perales Apaico
}
