% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piscomgroup.R
\name{piscomgroup}
\alias{piscomgroup}
\title{values extraction of PISCO monthly data from a stations group}
\usage{
piscomgroup(x, start = NULL, end = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO file name (in netCDF format), longitude and latitude from a stations group.}

\item{start}{OPTIONAL, start date of dataset, default value is 1981-01-01.}

\item{end}{OPTIONAL, end date of dataset, default value is 2016-12-01.}
}
\description{
function for extract values of PISCO monthly data from a stations group, PISCO is Peruvian Interpolated Data of the Senamhi’s Climatological and Hydrologycal Observations.
}
