% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpisco-package.R
\docType{package}
\name{rpisco-package}
\alias{rpisco}
\alias{rpisco-package}
\title{rpisco: Access and Process PISCO High-Resolution Climate Grids for Peru}
\description{
The \code{rpisco} package provides programmatic access, automated caching,
spatial subsetting, temporal analysis, and hydrological validation metrics
for the Peruvian Interpolated data of SENAMHI's Climatological and
hydrological Observations (PISCO).
}
\details{
Supports the complete family of PISCO datasets developed by SENAMHI DHI-SEH:
\itemize{
\item \strong{Precipitation (PISCOp)}: PISCOp v3.0 (1981-2025; daily, monthly, and
climatological normal) and PISCOp_h (1981-2023 coarse gridded rainfall).
\item \strong{Temperature (PISCOt)}: PISCOt v1.2 (1981-2016; maximum and minimum
daily, monthly, and climatological normal).
\item \strong{Evapotranspiration (PISCOeo_pm)}: PISCOeo_pm (1981-2016; daily,
monthly, and climatological normal reference evapotranspiration).
\item \strong{Rainfall Erosivity (PISCO_reed)}: PISCO_reed v1.0 (1981-2016; R-factor
climatology and annual erosivity series).
\item \strong{Streamflow (PISCO_HyM)}: Monthly gridded and vector streamflow
(1981-2016) based on hydrological modeling (GR2M and ARNOVIC) with sub-basin
polygons and river reach networks.
}
Datasets are hosted on Figshare and HydroShare by SENAMHI and collaborating
researchers. See \code{\link{pisco_catalog}} and \code{\link{pisco_citation}}
for links, dataset keys, and scientific citations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pefrens/rpisco}
  \item \url{https://pefrens.github.io/rpisco/}
  \item Report bugs at \url{https://github.com/pefrens/rpisco/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Leonardo Gutierrez (\href{https://orcid.org/0000-0002-8658-139X}{ORCID}) [data contributor]
  \item Waldo Sven Lavado-Casimiro [data contributor]
}

}
\keyword{internal}
