% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpisco-package.R
\docType{package}
\name{rpisco-package}
\alias{rpisco}
\alias{rpisco-package}
\title{rpisco: Access and Process PISCO High-Resolution Climate Grids for Peru}
\description{
The \code{rpisco} package provides programmatic access, automated caching,
spatial subsetting, temporal analysis, and hydrological validation metrics
for the Peruvian Interpolated data of SENAMHI's Climatological and
hydrological Observations (PISCO).
}
\details{
Specifically supports PISCOp v3.0 (1981-2025) gridded daily, monthly, and
normal climatology rainfall datasets published by Gutierrez and
Lavado-Casimiro (2025) on Figshare (\url{doi:10.6084/m9.figshare.32411886})
and documented in the official technical publication of the Servicio
Nacional de Meteorologia e Hidrologia del Peru (SENAMHI):
\url{https://hdl.handle.net/20.500.12542/4183}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pefrens/rpisco}
  \item \url{https://pefrens.github.io/rpisco/}
  \item Report bugs at \url{https://github.com/pefrens/rpisco/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefren16@gmail.com}

Authors:
\itemize{
  \item Paul Efren Santos Andrade \email{paulefren16@gmail.com}
}

Other contributors:
\itemize{
  \item Leonardo Gutierrez (\href{https://orcid.org/0000-0002-8658-139X}{ORCID}) [data contributor]
  \item Waldo Sven Lavado-Casimiro [data contributor]
}

}
\keyword{internal}
