% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyavg.R
\name{yearlyavg}
\alias{yearlyavg}
\title{Function for processing PISCO daily/monthly values into a yearly values}
\usage{
yearlyavg(x, param = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily/monthly values.}

\item{param}{OPTIONAL, default parameter is sum for precipitation and evapotranspiration,
use mean for temperature.}
}
\description{
Function for processing PISCO daily/monthly values into a yearly values, apply to precipitation,
temperature and evapotranspiration.
}
\examples{

library(RPisco)

pisco.ppd <- "D:/PISCOd_pp_2.0.nc"
latitude <- -74.23
longitude <- -13.18
x <-  data.frame(pisco.ppd, latitude, longitude)
data.diaria.pp <- piscod(x)
monthlyavg(data.diaria.pp)

}
\author{
Geomar Perales Apaico
}
