% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyavg.R
\name{yearlyavg}
\alias{yearlyavg}
\title{function for processing of pisco daily/monthly values to a yearly average.}
\usage{
yearlyavg(x, variable = NULL)
}
\arguments{
\item{x}{a dataframe with PISCO daily/monthly values.}

\item{variable}{OPTIONAL, default value is sum for precipitation, use mean for temperature.}
}
\description{
function for processing of pisco daily/monthly values to a yearly average, apply to precipitation and
temperature.
}
